requirement.txt
pandas==2.2.2
openpyxl==3.1.5
scikit-learn==1.5.1
tqdm==4.66.4
requests==2.32.3
python-dotenv==1.0.1
pyyaml==6.0.2
